refactor(ui): split TwitchTab into specialized subcomponents
- Created `TwitchAuthCard`, `TwitchEventsCard`, and `TwitchResponseLogicCard` to modularize the massive `TwitchTab.tsx`.
- Fixed React property typings in the new subcomponents.
- Improved overall maintainability of the settings tab.
